requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
transformers==4.35.0
torch>=2.2.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
googletrans==4.0.0rc1
requests>=2.31.0
nltk>=3.8.1
gunicorn>=21.2.0